backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask",
    "flask-socketio",
    "mblt-model-zoo[transformers]==1.0.0",
    "python-dotenv>=1.1.1",
]

[tool.black]
line-length = 120
